refactor(agent): split the launcher into options, supervisor, record mode and entry
Pure move: option parsing, the shared process supervisor and the
record mode each get their own module; the entry keeps standalone mode
and main. No behaviour change; the smoke and a record-mode recording
pass on the split.
